+++ packages/yourloops/lib/team/team.hook.tsx
@@ -54,9 +54,6 @@ import SpinningLoader from '../../components/loaders/spinning-loader'
 
 const ReactTeamContext = createContext<TeamContext>({} as TeamContext)
 
-export const PRIVATE_TEAM_ID = 'private'
-export const PRIVATE_TEAM_NAME = 'private'
-
 function TeamContextImpl(): TeamContext {
   const authHook = useAuth()
   const notificationHook = useNotification()
@@ -101,7 +98,7 @@ function TeamContextImpl(): TeamContext {
   }
 
   const getPrivateTeam = (): Team => {
-    return teams.find((team: Team) => TeamUtils.isPrivate(team))
+    return teams.find((team: Team) => TeamUtils.isPrivate(team.id))
   }
 
   const getTeamsByType = (type: TeamType): Team[] => {
